requirements.txt
control-factory
http-control
nodered-res
